+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Entity.relation_components` now has a MutableMapping functionality.
+- You can now set the value of `Entity.relation_components[component_key] = {target: component}`.
 
 ## [4.1.0] - 2023-07-28
 ### Added